- back to floats (except elliptical orbit) - fixed build (GraphicsSettings.asset)

+++ Assets/Scripts/FuturePhysics/FuturePhysics.cs
@@ -6,8 +6,8 @@ using System.Linq;
 public class FuturePhysics
 {
     public const int MaxSteps = 10000;
-    public const double DeltaTime = 0.04f;
-    public const double G = 1f;
+    public const float DeltaTime = 0.04f;
+    public const float G = 1f;
 
     public static int currentStep;
     public static int lastVirtualStep;
